refactoerd BookingDetailsPage into smaller components
