+++ frontend-web/src/components/BookingDetailsPage/index.ts
@@ -1,0 +1,7 @@
+export { default as BookingHeader } from './BookingHeader';
+export { default as BikeImageGallery } from './BikeImageGallery';
+export { default as BookingDetailsGrid } from './BookingDetailsGrid';
+export { default as RatingAndReviewSection } from './RatingAndReviewSection';
+export { default as QuickActions } from './QuickActions';
+export { default as BookingTimeline } from './BookingTimeline';
+export { default as AdditionalInfo } from './AdditionalInfo';
